backend_carpi/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.7.1
Django==5.1.1
django-cors-headers==4.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.10
#mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.4.0
pip==24.2
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==10.12.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
drf-yasg==1.21.7 #pip install --upgrade setuptools 